Add review refactoring 2
